Add support for data space invititations
Data space invititations consist of an existing owner within the
data space sending out an invite to any email. The invite token enables
the new user to broadcast their user identifier to the data space, or to
ignore the invite. Once accepted, any member of the data space has can
confirm the new user by first sharing the metadata key and then
announcing that the invitation flow is complete.

Finally, the invite token is also valid as a confirmation token for new
users. If it's valid, sending the confirmation mail is skipped and the
user is flagged as confirmed.